test: Remove duplicated code
